Move QueueableRealization and RealizationState into its own file
